+++ changelog.md
@@ -1,1 +1,1 @@
-random fixes
+- fixed a recent worldgen issue